Fix sorting of side-effect imports in Node <12
`Array.prototype.sort` in Node <12 does not use stable sorting, and this
plugin accidentally relied on stable sorting to keep the original order
of side-effect imports. This commit makes sure that side effect imports
end up in their original order regardless of whether `.sort` is stable
or not.

Fixes #34 and closes #36.